Move serializers into serializers module
